Make last_pushdata an inherent function
We are attempting to move the `Script` type to `primitives`, doing so
requires usage of the extension trait mechanism recently introduced.
Private methods cannot be created using traits only private traits. An
alternative to creating a private trait is to make a method into a
function and use the concrete as a parameter instead of taking `self`.

Make `Script::last_pushdata` an inherent, this is only an internal
change.