Separate private ScriptBuf methods
In preparation for adding a private `ScriptBufExtPriv` trait, move the
private methods to a separate impl block.

Make the `push_slice_no_op` method have the same visibility as the
others so that the change in scope does not get hidden in the upcoming
patch that introduces the trait.

(Note this is basically just a code move and the diff shows move of other
methods not the private ones.)