Introduce ScriptBuf extension traits
In preparation for moving the `ScriptBuf` type to the `primitives`
crate; introduce two `ScriptBuf` extension traits, one public and one
private.

Note, the private extension trait has `pub(crate)` where as before this
patch was applied we use `pub(in crate::script)`. This is because the
macro doesn't handle the latter syntax.

Note also, that the build failure when `ScriptBuf::from_hex` is not
found may be confusing because of the history of `from_hex` - users may
go looking for a `FromHex` trait. We should keep this in mind when
documenting how to use all the new extension traits.