psbt: Split serializable types out
Currently the `Psbt` data type is doing two things:

1. It is the BIP-174 PSBT abstraction, has fields named as in the bip
and does serialization/deserialization.

2. It provides a rich API e.g., signing.

We would like to split these two things apart so that we can start
supporting PSBTv2 (BIP-370) without rug pulling users of the API.

As a first step create some additional types, one set that supports
serilization and one set that supports the other API.

- `psbt::{Psbt, Input, Output}`: Rich API
- `psbt::map::{Psbt, Input, Output}`: Serialization

Provide conversion functions `into_serializable` and `from_serializable`
for all types. Note that it would be nice to implement `as_serializable`
also because the serialize API does not consume self but conversion to a
serializable type does, currently we use a call to `clone()`.