Manually format the codebase
When we messed with the re-exports recently we failed to notice that
the formatter would move the `pub use` blocks around because I only
put `rustfmt::skip` on one block, I thought that had the effect of
stopping all exports below that block being changed, turns out I was
wrong.

Run the formatter then make some minor changes to the re-exports in
`bitcoin`, including changing comments that do not make sense with the
new layout.