Enforce MAX_MONEY invariant in amount types
This is a complete audit of the `amount` module.

Recently we changed MAX to equal MAX_MONEY for the two amount types. In
doing so we left the module in quite a mess.

Note also we are trying to 1.0 `units` so this patch goes over the docs
with a fine tooth comb. Note also please that in order to catch mistakes
we are going to quite some effort to make the unsigned and signed files
similar, this patch pushes that further by using `Self::` in both docs
and code instead of the actual amount type.

Enforcing the `MAX_MONEY` invariant is quite involved because it means
multiple things:

- Constructing amounts is now fallible
- Converting from unsigned to signed is now infallible
- Taking the absolute value is now infallible
- Integer overflow is illuminated in various places

Details:

- Update `from_sat` to check the invariant
- Fix all docs including examples
- Use the unchecked constructor in test code
- Comment any other use of the unchecked constructor
- Comment any place we do unchecked integer ops
- Deprecate `unchecked_abs`
- Fail serde using the horrible string error
- Try not to use the unchecked constructor in rustdocs, no need to
  encourage unsuspecting users to use it.
- Fix the txout regression test to use MAX_MONEY
- Use unchecked constructor in `TxOut::NULL`
- Remove `iter::Sum` impls
- Favour `MAX_MONEY` in code over `MAX` just to be explicit and save
  unsuspecting devs from missing this change.
- Use `?` in rustdoc examples (required by Rust API guidlines)
- Add `# Returns` section to checked ops functions, this differs from
  other places in the crate because the amount types differ also so I
  believe the non-uniformity is warranted.
- Remove `TryFrom<Amount> for SignedAmount` because the conversion is
- now infallible. Add a `From` impl.
- Fix the arbitrary impls
- Maintain correct formatting
- Add units test that verify overflow assumptions (these are testing
  Rust really but I was paranoid about introducing bugs)
- Remove private `check_max` function as its no longer needed