Remove marker traits from network validation traits
In PR #1785 it was claimed that we needed marker traits in order for
`Serialize` to be derived because of the `NetworkValidation` trait
bound.

I can't work out why. Note I ack'd the PR so that is my bad.

We have a unit test that derives `Serialize` on a type that includes an
address with generic trait bound. So the fact that unit tests pass with
this patch applied shows we do not need the marker trait bounds - I
think.